fix: Product Models type

+++ netcore-webapi-practice/Models/Product.cs
@@ -7,10 +7,10 @@
         public int SupplierID { get; set; }
         public int CategoryID { get; set; }
         public string? QuantityPerUnit { get; set; }
-        public double UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
-        public int UnitsOnOrder { get; set; }
-        public int ReorderLevel { get; set; }
+        public decimal UnitPrice { get; set; }
+        public short UnitsInStock { get; set; }
+        public short UnitsOnOrder { get; set; }
+        public short ReorderLevel { get; set; }
         public bool Discontinued { get; set; }
     }
 }
